Modify the first assignment

+++ src/main/java/com/example/AEPB/AirCoin.java
@@ -1,6 +1,8 @@
 package com.example.AEPB;
 
+import java.math.BigDecimal;
+
 public class AirCoin {
 
-    Double amount;
+    BigDecimal amount;
 }
