+++ src/test/java/com/example/AEPB/AepbApplicationTests.java
@@ -3,6 +3,8 @@ package com.example.AEPB;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.assertTrue;
 
 class HelloTest  {
@@ -11,9 +13,9 @@ class HelloTest  {
     void should_equals_when_compare_two_airCoin_amount_given_amount_is_11_and_amount_is_11() {
         //given
         AirCoin AirCoin1 = new AirCoin();
-        AirCoin1.amount = 11.0;
+        AirCoin1.amount = BigDecimal.valueOf(11.0);
         AirCoin AirCoin2 = new AirCoin();
-        AirCoin2.amount = 11.0;
+        AirCoin2.amount = BigDecimal.valueOf(11.0);
 
         //when
         Boolean b = AmountUtils.compareTwoAirCoinAmount(AirCoin1, AirCoin2);
@@ -26,9 +28,9 @@ class HelloTest  {
     void should_not_equals_when_compare_two_airCoin_amount_given_amount_is_11_dot_0_and_amount_is_12_dot_0() {
         //given
         AirCoin AirCoin1 = new AirCoin();
-        AirCoin1.amount = 11.0;
+        AirCoin1.amount = BigDecimal.valueOf(11.0);
         AirCoin AirCoin2 = new AirCoin();
-        AirCoin2.amount = 12.0;
+        AirCoin2.amount = BigDecimal.valueOf(12.0);
 
         //when
         Boolean b = AmountUtils.compareTwoAirCoinAmount(AirCoin1, AirCoin2);
@@ -41,9 +43,9 @@ class HelloTest  {
     public void should_throw_exception_when_amount_is_11_and_amount_999999999999999999999999999() {
         //given
         AirCoin AirCoin1 = new AirCoin();
-        AirCoin1.amount = 11.0;
+        AirCoin1.amount = BigDecimal.valueOf(11.0);
         AirCoin AirCoin2 = new AirCoin();
-        AirCoin2.amount = 999999999999999999999999999.0;
+        AirCoin2.amount = BigDecimal.valueOf(999999999999999999999999999.0);
 
         // when
         Throwable t = null;
